Calculates atmospheric pressure

Calculates atmospheric pressure as a function of elevation, by default assuming standard atmosphere (101325 Pa at sea level)

calc_patm(elv, patm0 = 101325)

Arguments

elv
Elevation above sea-level (m.a.s.l.)

patm0
(Optional) Atmospheric pressure at sea level (Pa), defaults to 101325 Pa.

Value

A numeric value for \(p\)

Details

The elevation-dependence of atmospheric pressure is computed by assuming a linear decrease in temperature with elevation and a mean adiabatic lapse rate (Berberan-Santos et al., 1997): $$ p(z) = p0 ( 1 - Lz / TK0) ^ ( g M / (RL) ) $$ where \(z\) is the elevation above mean sea level (m, argument elv), \(g\) is the gravity constant (9.80665 m s-2), \(p0\) is the atmospheric pressure at 0 m a.s.l. (argument patm0, defaults to 101325 Pa), \(L\) is the mean adiabatic lapse rate (0.0065 K m-2), \(M\) is the molecular weight for dry air (0.028963 kg mol-1), \(R\) is the universal gas constant (8.3145 J mol-1 K-1), and \(TK0\) is the standard temperature (298.15 K, corresponds to 25 deg C).

References

Allen, R. G., Pereira, L. S., Raes, D., Smith, M.: FAO Irrigation and Drainage Paper No. 56, Food and Agriculture Organization of the United Nations, 1998

Examples

print("Standard atmospheric pressure, in Pa, corrected for 1000 m.a.s.l.:")

#> [1] "Standard atmospheric pressure, in Pa, corrected for 1000 m.a.s.l.:"

print(calc_patm(1000))

#> [1] 90241.54