Defines settings for settings for FLUXNET data ingestion

Defines settings for settings for FLUXNET data ingestion

get_settings_fluxnet(
  dir_hh = ".",
  dir_hr = ".",
  getswc = TRUE,
  threshold_GPP = 0,
  threshold_LE = 0,
  threshold_H = 0,
  threshold_SWC = 0,
  threshold_WS = 0,
  threshold_USTAR = 0,
  threshold_T = 0,
  threshold_NETRAD = 0,
  filter_ntdt = FALSE,
  return_qc = FALSE,
  remove_neg = FALSE,
  verbose = TRUE
)

Arguments

dir_hh

A character string specifying the local path of half-hourly FLUXNET data, required to get daytime VPD. Defaults to "." (present working directory).

dir_hr

A character string specifying the local path of hourly FLUXNET data, required to get daytime VPD. Defaults to "." (present working directory).

getswc

If getswc==TRUE, then all soil water content data (variables starting with "SWC_") are read. Defaults to TRUE.

threshold_GPP

A numeric value (between 0 and 1 for daily, weekly, monthly, and annual data or 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data). The value specifies the threshold for excluding data during data cleaning. The threshold is with respect to the data quality flag in the FLUXNET data, indicating the fraction of measured and good quality gapfilled data for daily, weekly, monthly, and annual data and 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data. Defaults to threshold_GPP=0 meaning no data is excluded.

threshold_LE

A numeric value (between 0 and 1 for daily, weekly, monthly, and annual data or 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data). The value specifies the threshold for excluding data during data cleaning. The threshold is with respect to the data quality flag in the FLUXNET data, indicating the fraction of measured and good quality gapfilled data for daily, weekly, monthly, and annual data and 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data. Defaults to threshold_GPP=0 meaning no data is excluded.

threshold_H

A numeric value (between 0 and 1 for daily, weekly, monthly, and annual data or 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data). The value specifies the threshold for excluding data during data cleaning. The threshold is with respect to the data quality flag in the FLUXNET data, indicating the fraction of measured and good quality gapfilled data for daily, weekly, monthly, and annual data and 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data. Defaults to threshold_GPP=0 meaning no data is excluded.

threshold_SWC

A numeric value (between 0 and 1 for daily, weekly, monthly, and annual data or 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data). The value specifies the threshold for excluding data during data cleaning. The threshold is with respect to the data quality flag in the FLUXNET data, indicating the fraction of measured and good quality gapfilled data for daily, weekly, monthly, and annual data and 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data. Defaults to threshold_GPP=0 meaning no data is excluded.

threshold_WS

A numeric value (between 0 and 1 for daily, weekly, monthly, and annual data or 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data). The value specifies the threshold for excluding data during data cleaning. The threshold is with respect to the data quality flag in the FLUXNET data, indicating the fraction of measured and good quality gapfilled data for daily, weekly, monthly, and annual data and 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data. Defaults to threshold_GPP=0 meaning no data is excluded.

threshold_USTAR

A numeric value (between 0 and 1 for daily, weekly, monthly, and annual data or 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data). The value specifies the threshold for excluding data during data cleaning. The threshold is with respect to the data quality flag in the FLUXNET data, indicating the fraction of measured and good quality gapfilled data for daily, weekly, monthly, and annual data and 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data. Defaults to threshold_GPP=0 meaning no data is excluded.

threshold_T

A numeric value (between 0 and 1 for daily, weekly, monthly, and annual data or 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data). The value specifies the threshold for excluding data during data cleaning. The threshold is with respect to the data quality flag in the FLUXNET data, indicating the fraction of measured and good quality gapfilled data for daily, weekly, monthly, and annual data and 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data. Defaults to threshold_GPP=0 meaning no data is excluded.

threshold_NETRAD

A numeric value (between 0 and 1 for daily, weekly, monthly, and annual data or 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data). The value specifies the threshold for excluding data during data cleaning. The threshold is with respect to the data quality flag in the FLUXNET data, indicating the fraction of measured and good quality gapfilled data for daily, weekly, monthly, and annual data and 0 [measured], 1 [good quality gapfill], 2 [ medium], 3 [poor] for half-hourly data. Defaults to threshold_GPP=0 meaning no data is excluded.

filter_ntdt

A logical specifying whether agreement of daytime and nighttime based GPP estimates is to be used as a filter. Data points are removed where their difference is below the the 97.5 above the 2.5 Defaults to FALSE.

return_qc

A logical specifying whether quality control variables should be returned. Defaults to FALSE.

remove_neg

A logical specifying whether negative GPP values are to be removed (replaces with NA). Defaults to FALSE.

verbose

verbose output

Value

A named list containing information required for read data from standard FLUXNET data files (CSV files).

Examples

if (FALSE) settings_gee <- get_settings_fluxnet() # \dontrun{}