Defines settings for settings for Google Earth Engine download

Defines settings for settings for Google Earth Engine download for a pre-defined set of "bundles" (c("modis_fpar", "modis_evi", "modis_lai", "modis_gpp")).

get_settings_gee(
  bundle = "modis_fpar",
  python_path = system("which python", intern = TRUE),
  gee_path = ".",
  data_path = ".",
  method_interpol = "linear",
  keep = FALSE,
  overwrite_raw = FALSE,
  overwrite_interpol = FALSE
)

Arguments

bundle

A character string specifying which dataset (bundle) to download. Defaults to "modis_fpar". Available are: c("modis_fpar", "modis_evi", "modis_lai", "modis_gpp").

python_path

A character string specifying the local path to the python executable

gee_path

A character string specifying the local path to the gee_subseet library. Defaults to "." (present working directory).

data_path

A character string specifying the path of where the data should be downloaded to. Defaults to "." (present working directory).

method_interpol

A character string specifying which interpolation method to use. Defaults to linear interpolation ("linear").

keep

A logical specifying whether to keep all intermediate data (before filtering, and before imputing mean seasonal cycle), and all alternative interpolation results. Defaults to FALSE.

overwrite_raw

A logical specifying whether raw data as downloaded from GEE is to be overwritten. Defaults to FALSE, i.e. data is read from exisitng file if available.

overwrite_interpol

A logical specifying whether processed (interpolated) data, is to be overwritten. Defaults to FALSE, i.e. data is read from exisitng file if available.

Value

A named list containing information required for download from Google Earth Engine.

Examples

if (FALSE) settings_gee <- get_settings_gee() # \dontrun{}