Defines settings for settings for MODIS download using MODISTools

Defines settings for settings for Google Earth Engine download for a pre-defined set of "bundles" (c("modis_fpar", "modis_evi", "modis_lai", "modis_gpp")).

get_settings_modis(
  bundle = "modis_fpar",
  data_path = ".",
  method_interpol = "linear",
  keep = FALSE,
  overwrite_raw = FALSE,
  overwrite_interpol = FALSE,
  n_focal = 0,
  filename_with_year = TRUE,
  network = NA
)

Arguments

bundle
A character string specifying which dataset (bundle) to download. Defaults to "modis_fpar". Available are: c("modis_fpar", "modis_evi", "modis_lai", "modis_gpp").

data_path
A character string specifying the path of where the data should be downloaded to. Defaults to "." (present working directory).

method_interpol
A character string specifying which interpolation method to use. Defaults to linear interpolation ("linear"). Alternatives are

keep
A logical specifying whether to keep all intermediate data (before filtering, and before imputing mean seasonal cycle), and all alternative interpolation results. Defaults to FALSE.

overwrite_raw
A logical specifying whether raw data as downloaded from GEE is to be overwritten. Defaults to FALSE, i.e. data is read from exisitng file if available.

overwrite_interpol
A logical specifying whether processed (interpolated) data, is to be overwritten. Defaults to FALSE, i.e. data is read from exisitng file if available.

n_focal
An integer specifying the distance (in number of pixels) around the center pixel to be used for averaging. Defaults to zero (using only the center pixel).

filename_with_year
A logical specifying whether the years covered are specified in the file name. Added here for consistency with earlier versions of ingestr where years were not specified.

network
A character string specifying the network for which the site names correspond. Earth Engine.

Examples

if (FALSE) settings_gee <- get_settings_gee()