Ingest Pacific Decadal Oscillation data

ingest_PDO ingests data from the updated standardized values for the PDO index, derived as the leading PC of monthly SST anomalies in the North Pacific Ocean, poleward of 20N. The monthly mean global average SST anomalies are removed to separate this pattern of variability from any "global warming" signal that may be present in the data. If you have any questions about this time series, contact Nathan Mantua at: nate.mantua\@noaa.gov All ingest functions use the source file name as an identifying column to track provenance and relate data and metadata read from files.

ingest_PDO(input.source = "http://jisao.washington.edu/pdo/PDO.latest",
  end.year = NULL, export.header = TRUE)

Arguments

input.source
Character indicating the URI to the HTML representation of the data.

end.year
Four digit integer indicating the last year of data wanted.

export.header
A logical indicating if header information is written to a separate data frame.

Value

A data frame. If export.header = TRUE a temporary file is created for the header data. See ingest_header for more information.

Examples

df_pdo <- ingest_PDO()  # reads in all the data from start date to present

#> Header info for http://jisao.washington.edu/pdo/PDO.latest has been saved to a temporary file. Run ingest_header('http://jisao.washington.edu/pdo/PDO.latest') to load the header data.

df_pdo1 <- ingest_PDO(end.year=2000)  # reads in the data from start date to the year 2000

#> Header info for http://jisao.washington.edu/pdo/PDO.latest has been saved to a temporary file. Run ingest_header('http://jisao.washington.edu/pdo/PDO.latest') to load the header data.