Defines settings for settings for Google Earth Engine download

Returns a list of two data frames, one with data at original modis dates (df), and one interpolated to all days (ddf).

ingest_gee_bysite(
  df_siteinfo,
  start_date,
  end_date,
  overwrite_raw,
  overwrite_interpol,
  band_var,
  band_qc,
  prod,
  prod_suffix,
  varnam,
  productnam,
  scale_factor,
  period,
  python_path,
  gee_path,
  data_path,
  method_interpol,
  keep
)

Arguments

df_siteinfo
xxx

start_date
xxx

end_date
xxx

overwrite_raw
xxx

overwrite_interpol
xxx

band_var
xxx

band_qc
xxx

prod
xxx

prod_suffix
xxx

varnam
xxx

productnam
xxx

scale_factor
xxx

period
xxx

python_path
xxx

gee_path
xxx

data_path
xxx

method_interpol
xxx

keep
xxx

Value

A named list containing information required for download from Google Earth Engine.

Examples

settings_gee <- get_settings_gee( bundle = "modis_fpar" )