Complements the setup settings

Complements the settings based on the site meta info CSV file or data frame.

prepare_setup_sofun(site_info, params_siml)

Arguments

site_info

A character string specifying the path to the site meta information file, or a data frame containing the site meta info. Required columns are:

sitename

Name of the site, must be the first column of the file.

lon

Longitude of site.

lat

Latitude of site.

elv

Elevation of site, in m.

year_start, year_end

Years for which the simulation is to be done, corresponding to data availability from site.

params_siml

A named list containing the simulation parameters for SOFUN.

Value

A data frame (tibble) containing the site meta information, complemented by column params_siml which is a nested list of complemented simulation parameters.

Examples

if (FALSE) { # \dontrun{
 setup <- prepare_setup_sofun(
   site_info = site_info,
   params_siml = params_siml
  )
} # }