Data ingest for a single site

Ingests data for a single site and one specific data type, specified by argument source.

ingest_bysite(
  sitename,
  source,
  getvars = NULL,
  dir = NULL,
  settings = list(),
  timescale = "d",
  year_start,
  year_end,
  lon = ifelse(source == "fluxnet2015", NA),
  lat = ifelse(source == "fluxnet2015", NA),
  verbose = FALSE
)

Arguments

sitename
A character string used as site identification. When data is extracted from global files or remote servers, sitename is simply used as a label and any string can be provided. When data is extraced from site-specific files (e.g. source = "FLUXNET2015"), then sitename is used to identify the file from which data is read.

source
A character used as identifiyer for the type of data source (e.g., "fluxnet2015").

getvars
A named list of characters specifying the variable names in the source dataset corresponding to standard names "temp" for temperature, "prec" for precipitation, "patm" for atmospheric pressure, "vpd" for vapour pressure deficit, "netrad" for net radiation, "swin" for shortwave incoming radiation.

dir
A character specifying the directory where data is located.

settings
A list of additional settings used for reading original files.

timescale
A character or vector of characters, specifying the time scale of data used from the respective source (if multiple time scales are available, otherwise is disregarded).

year_start
An integer specifying the first year for which data is to be ingested.

year_end
An integer specifying the last year for which data is to be ingested (full years are read, i.e. all days, or hours, or months in each year).

lon
A numeric value specifying the longitude for which data is extraced from global files or remote data servers. If source = "FLUXNET2015", this is not required and set ot NA.

lat
A numeric value specifying the longitude for which data is extraced from global files or remote data servers. If source = "FLUXNET2015", this is not required and set ot NA.

verbose
if TRUE, additional messages are printed.

Value

A data frame (tibble) containing the time series of ingested data.

Examples

if (FALSE) inputdata <- ingest_bysite()